+++ src/ConfBase.h
@@ -53,6 +53,7 @@ public:
       pruneParameters(parent->pruneParameters), preVolume(parent->preVolume),
       postVolume(parent->postVolume),
       backupJobTimeout(parent->backupJobTimeout),
+      rsyncIOTimeout(parent->rsyncIOTimeout),
       rsyncCommand(parent->rsyncCommand),
       rsyncBaseOptions(parent->rsyncBaseOptions),
       rsyncExtraOptions(parent->rsyncExtraOptions),
@@ -87,6 +88,9 @@ public:
 
   /** @brief backup job timeout */
   int backupJobTimeout = 0;
+
+  /** @brief rsync IO timeout */
+  int rsyncIOTimeout = 0;
 
   /** @brief rsync command */
   std::string rsyncCommand = "rsync";